improve plugin load
